requirements.txt
ansible==2.7.10
appdirs==1.4.3
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
decorator==4.4.0
dogpile.cache==0.7.1
idna==2.8
iso8601==0.1.12
Jinja2==2.10.1
jmespath==0.9.4
jsonpatch==1.23
jsonpointer==2.0
keystoneauth1==3.14.0
MarkupSafe==1.1.1
munch==2.3.2
netifaces==0.10.9
openstacksdk==0.28.0
os-client-config==1.32.0
os-service-types==1.7.0
paramiko==2.4.2
pbr==5.2.0
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
PyYAML==5.1
requests==2.22.0
requestsexceptions==1.4.0
shade==1.31.0
six==1.12.0
stevedore==1.30.1
texttable==1.6.1
urllib3==1.25.2
